use bcrypt to hash password during signup
The session-based authorization is implement on a superficial level. The auth.go file contains TODOs. The session life-cycle is limited to a single day for now, meaning that players will have to sign in on a daily basis. This limitation isn't a major issue during the current development phase, but is should be addressed as soon as possible.